Disable DHCP service
Our main use cases for Bananapro do not require DHCP service to be active, in fact
it causes problems when attached to a network with existing DHCP service.

FIXME: probably could find a better way to do this, maybe as a menuconfig option
that triggers a patch to be applied? (Needs research)